Created Node<T,I>
In Node<T,I>, the node doesn't have to be a simple string, like in Node.
It can be any object of type T that implements IIdentifiable<I>, i.e.,
has a property Id of type I.
Unit tests test for nodes with string or int as Id. The second one has
some throwaway properties (a double and an integer), just to add some
spice.